+++ Koronavilkku/Koronavilkku/ViewController/TestViewController.swift
@@ -24,11 +24,13 @@ class TestViewController: UIViewController {
     lazy var setLastExposureCheckButton = self.createButton(title: "Set exposure check date", action: #selector(setLastExposureCheck))
     lazy var dumpEFGSCountriesButton = self.createButton(title: "Dump EFGS countries", action: #selector(dumpEFGSCountries))
     lazy var deleteEFGSCountriesButton = self.createButton(title: "Delete EFGS countries", action: #selector(deleteEFGSCountries))
+    lazy var showNotificationButton = self.createButton(title: "Show delayed notification", action: #selector(showDelayedNotification))
 
     // force downcast, because we're using the internals here
     var batchRepository = Environment.default.batchRepository as! BatchRepositoryImpl
     var exposureManager = ExposureManagerProvider.shared.manager
-    var exposureRepository = Environment.default.exposureRepository
+    var exposureRepository = Environment.default.exposureRepository as! ExposureRepositoryImpl
+    var notificationService = Environment.default.notificationService
     
     override func viewDidLoad() {
         super.viewDidLoad()
@@ -39,29 +41,46 @@ class TestViewController: UIViewController {
         
         self.view.addKeyboardDisposer()
         
-        view.addSubview(batchIdInput)
+        let scrollView = UIScrollView()
+        view.addSubview(scrollView)
+        
+        scrollView.snp.makeConstraints { make in
+            make.top.bottom.equalTo(view.safeAreaInsets)
+            make.left.right.equalToSuperview()
+        }
+        
+        let content = UIView()
+        scrollView.addSubview(content)
+        
+        content.snp.makeConstraints { make in
+            make.edges.equalTo(scrollView.contentLayoutGuide)
+            make.width.equalTo(scrollView.frameLayoutGuide)
+        }
+        
+        content.addSubview(batchIdInput)
         batchIdInput.placeholder = "Batch id"
         batchIdInput.borderStyle = .bezel
         batchIdInput.textColor = .black
         batchIdInput.snp.makeConstraints { make in
             make.width.equalTo(150)
             make.height.equalTo(30)
-            make.right.equalTo(view.snp.centerX)
-            make.top.equalTo(view.safeAreaLayoutGuide.snp.top).offset(50)
+            make.right.equalTo(content.snp.centerX)
+            make.top.equalToSuperview().offset(50)
         }
         
-        view.addSubview(storeBatchIdButton)
+        content.addSubview(storeBatchIdButton)
+        
         self.storeBatchIdButton.snp.makeConstraints { make in
             make.width.equalTo(150)
             make.height.equalTo(30)
             make.left.equalTo(batchIdInput.snp.right).offset(10)
-            make.top.equalTo(view.safeAreaLayoutGuide.snp.top).offset(50)
+            make.top.equalToSuperview().offset(50)
         }
         
         var top = batchIdInput.snp.bottom
         
         func appendButton(_ button: UIButton) {
-            self.view.addSubview(button)
+            content.addSubview(button)
             button.snp.makeConstraints { make in
                 make.width.equalTo(200)
                 make.height.equalTo(30)
@@ -84,6 +103,11 @@ class TestViewController: UIViewController {
         appendButton(setLastExposureCheckButton)
         appendButton(dumpEFGSCountriesButton)
         appendButton(deleteEFGSCountriesButton)
+        appendButton(showNotificationButton)
+        
+        scrollView.snp.makeConstraints { make in
+            make.bottom.equalTo(top).offset(30)
+        }
 
         LocalStore.shared.$uiStatus.addObserver(using: { [weak self] in
             self?.radarStatus.setTitle("Radar status \(LocalStore.shared.uiStatus)", for: .normal)
@@ -148,11 +172,13 @@ class TestViewController: UIViewController {
         
         LocalStore.shared.exposureNotifications.append(notification)
         LocalStore.shared.updateDateLastPerformedExposureDetection()
+        exposureRepository.showExposureNotification(delay: nil)
         Log.d("Created exposure notification \(notification)")
     }
     
     @objc func removeExposures() {
         LocalStore.shared.resetExposures()
+        notificationService.hideBadge()
     }
     
     @objc func toggleRadarStatus() {
@@ -214,6 +240,10 @@ class TestViewController: UIViewController {
         FileStorageImpl().delete(filename: EFGSRepositoryImpl.countryListFile)
     }
 
+    @objc func showDelayedNotification() {
+        exposureRepository.showExposureNotification(delay: 5)
+    }
+    
     private func showDialog(_ message: String, title: String = "Error") {
         showAlert(title: title, message: message, buttonText: "Dismiss")
     }
@@ -230,6 +260,8 @@ class TestViewController: UIViewController {
         case .btOff:
             return .apiDisabled
         case .apiDisabled:
+            return .notificationsOff
+        case .notificationsOff:
             return .on
         }
     }